added confirmation before deleting note
git-svn-id: http://svn.todd.net/svn/biblestudy/trunk@73 d5caad04-189b-467d-8c69-358c65116b6f

+++ net.todd.biblestudy.rcp/src/net/todd/biblestudy/rcp/presenters/NotePresenter.java
@@ -255,9 +255,12 @@ public class NotePresenter implements INoteListener, ICreateLinkListener
 
 	private void handleNoteDeleted()
 	{
-		String secondaryId = noteModel.getNote().getName();
-		noteModel.deleteNoteAndLinks();
-		noteView.closeView(secondaryId);
+		if (noteView.openDeleteConfirmationWindow() == 1)
+		{
+			String secondaryId = noteModel.getNote().getName();
+			noteModel.deleteNoteAndLinks();
+			noteView.closeView(secondaryId);
+		}
 	}
 
 	private void handleSaveNote()
@@ -333,7 +336,7 @@ public class NotePresenter implements INoteListener, ICreateLinkListener
 		}
 	}
 
-	private void handleOpenNote()
+	void handleOpenNote()
 	{
 		noteView.setViewTitle(noteModel.getNote().getName());
 		noteView.setContentText(noteModel.getNote().getText());
